+++ Changelog.md
@@ -10,6 +10,7 @@
 #### Improvement
 
 * amibo, juosan, mashu, starbattle: Reword English error messages
+* ayeheya: Replace check for rectangular room by check for symmetric room
 * cbblock: Add error for dead-end walls
 * heyawake: Don't autocheck until the grid is fully decided
 * heyawake: Don't insist on rectangular rooms, and treat